+++ src/regularboard.cc
@@ -2,16 +2,16 @@
 #include <algorithm>
 
 namespace nsweeper {
-RegularBoard::RegularBoard(size_t width, size_t height)
-    : Board{width * height}, dim{width, height} {
+RegularBoard::RegularBoard(int width, int height)
+    : Board{static_cast<size_t>(width * height)}, dim{width, height} {
   // Connect adjacent cells
-  for (size_t y = 0; y < getHeight(); ++y) {
-    for (size_t x = 0; x < getWidth(); ++x) {
+  for (int y = 0; y < getHeight(); ++y) {
+    for (int x = 0; x < getWidth(); ++x) {
       auto &cell = getCell(x, y);
-      for (size_t ny = std::max(y, static_cast<size_t>(1)) - 1;
-           ny <= std::min(y, getHeight() - 2) + 1; ++ny) {
-        for (size_t nx = std::max(x, static_cast<size_t>(1)) - 1;
-             nx <= std::min(x, getWidth() - 2) + 1; ++nx) {
+      for (int ny = std::max(y, 1) - 1; ny <= std::min(y, getHeight() - 2) + 1;
+           ++ny) {
+        for (int nx = std::max(x, 1) - 1; nx <= std::min(x, getWidth() - 2) + 1;
+             ++nx) {
           cell.addAdjacentCell(getCell(nx, ny));
         }
       }
@@ -19,11 +19,11 @@ RegularBoard::RegularBoard(size_t width, size_t height)
   }
 }
 
-size_t RegularBoard::getWidth() const { return dim.x; }
+int RegularBoard::getWidth() const { return dim.x; }
 
-size_t RegularBoard::getHeight() const { return dim.y; }
+int RegularBoard::getHeight() const { return dim.y; }
 
 size_t RegularBoard::toIndex(Vec2 pos) const {
-  return pos.y * getWidth() + pos.x;
+  return static_cast<size_t>(pos.y * getWidth() + pos.x);
 }
 } // namespace nsweeper
